Move BCrypt class to avoid conflict with org.mindrot:jbcrypt
- Renamed BCryptTest and updated using JUnit Test annotations
