+++ src/main/java/com/hierynomus/sshj/userauth/keyprovider/bcrypt/BCrypt.java
@@ -12,7 +12,7 @@
 // ACTION OF CONTRACT, NEGLIGENCE OR OTHER TORTIOUS ACTION, ARISING OUT OF
 // OR IN CONNECTION WITH THE USE OR PERFORMANCE OF THIS SOFTWARE.
 
-package org.mindrot.jbcrypt;
+package com.hierynomus.sshj.userauth.keyprovider.bcrypt;
 
 import java.io.UnsupportedEncodingException;
 import java.security.DigestException;
